Relocated classes in packages.
The spaceship pacakage was removed. The route classes in there are no in their own pacakage and the spaceship class belongs in the main package.
