+++ src/main/java/pixelperfect/route/RouteGenerator.java
@@ -1,4 +1,4 @@
-package pixelperfect.spaceship;
+package pixelperfect.route;
 
 /**
  * Random generator for routes through space.
